Sidebar: refactor map settings tab content refs: ZN-62
